Renamed request to gist_request to
prevent collision with other modules
